Rename src to asset_forwarder
